Move txAppender to separate source file
Code of both txAppender and stateManager is overcomplicated and must be simplified.
Separating is first step on this path.
