+++ pkg/state/state.go
@@ -16,7 +16,6 @@ import (
 	"github.com/wavesplatform/gowaves/pkg/keyvalue"
 	"github.com/wavesplatform/gowaves/pkg/proto"
 	"github.com/wavesplatform/gowaves/pkg/settings"
-	"github.com/wavesplatform/gowaves/pkg/types"
 	"github.com/wavesplatform/gowaves/pkg/util/lock"
 	"go.uber.org/zap"
 )
@@ -143,555 +142,6 @@ func (s *blockchainEntitiesStorage) flush(initialisation bool) error {
 		return err
 	}
 	return nil
-}
-
-type txAppender struct {
-	state types.SmartState
-	sc    *scriptCaller
-	ia    *invokeApplier
-
-	rw *blockReadWriter
-
-	stor     *blockchainEntitiesStorage
-	settings *settings.BlockchainSettings
-
-	// TransactionHandler is handler for any operations on transactions.
-	txHandler *transactionHandler
-	// Block differ is used to create diffs from blocks.
-	blockDiffer *blockDiffer
-	// Storage for diffs of incoming transactions (from added blocks or UTX).
-	// It will be used for validation and applying diffs to existing balances.
-	diffStor *diffStorage
-	// diffStorInvoke is storage for partial diffs generated by Invoke transactions.
-	// It is used to calculate balances that take into account intermediate invoke changes for RIDE.
-	diffStorInvoke *diffStorageWrapped
-	// Ids of all transactions whose diffs are currently in diffStor.
-	// This is needed to check that transaction ids are unique.
-	recentTxIds map[string]struct{}
-	// diffApplier is used to both validate and apply balance diffs.
-	diffApplier *diffApplier
-
-	// totalScriptsRuns counts script runs for UTX validation.
-	// It is increased every time ValidateNextTx() is called with transaction
-	// that involed calling scripts.
-	totalScriptsRuns uint64
-}
-
-func newTxAppender(
-	state types.SmartState,
-	rw *blockReadWriter,
-	stor *blockchainEntitiesStorage,
-	settings *settings.BlockchainSettings,
-) (*txAppender, error) {
-	sc, err := newScriptCaller(state, stor, settings)
-	if err != nil {
-		return nil, err
-	}
-	genesis, err := settings.GenesisGetter.Get()
-	if err != nil {
-		return nil, err
-	}
-	txHandler, err := newTransactionHandler(genesis.BlockSignature, stor, settings)
-	if err != nil {
-		return nil, err
-	}
-	blockDiffer, err := newBlockDiffer(txHandler, stor, settings)
-	if err != nil {
-		return nil, err
-	}
-	diffStor, err := newDiffStorage()
-	if err != nil {
-		return nil, err
-	}
-	diffStorInvoke, err := newDiffStorageWrapped(diffStor)
-	if err != nil {
-		return nil, err
-	}
-	diffApplier, err := newDiffApplier(stor.balances)
-	if err != nil {
-		return nil, err
-	}
-	ia := newInvokeApplier(state, sc, txHandler, stor, settings, blockDiffer, diffStorInvoke, diffApplier)
-	return &txAppender{
-		state:          state,
-		sc:             sc,
-		ia:             ia,
-		rw:             rw,
-		stor:           stor,
-		settings:       settings,
-		txHandler:      txHandler,
-		blockDiffer:    blockDiffer,
-		recentTxIds:    make(map[string]struct{}),
-		diffStor:       diffStor,
-		diffStorInvoke: diffStorInvoke,
-		diffApplier:    diffApplier,
-	}, nil
-}
-
-func (a *txAppender) checkDuplicateTxIdsImpl(id []byte, recentIds map[string]struct{}) error {
-	// Check recent.
-	if _, ok := recentIds[string(id)]; ok {
-		return errors.Errorf("transaction with ID %v already in state", id)
-	}
-	// Check DB.
-	if _, err := a.rw.readTransaction(id); err == nil {
-		return errors.Errorf("transaction with ID %v already in state", id)
-	}
-	return nil
-}
-
-func (a *txAppender) checkDuplicateTxIds(tx proto.Transaction, recentIds map[string]struct{}, timestamp uint64) error {
-	if tx.GetTypeVersion().Type == proto.PaymentTransaction {
-		// Payment transactions are deprecated.
-		return nil
-	}
-	if tx.GetTypeVersion().Type == proto.CreateAliasTransaction {
-		if (timestamp >= a.settings.StolenAliasesWindowTimeStart) && (timestamp <= a.settings.StolenAliasesWindowTimeEnd) {
-			// At this period alias transactions might have duplicate IDs due to bugs in historical blockchain.
-			return nil
-		}
-	}
-	txID, err := tx.GetID()
-	if err != nil {
-		return err
-	}
-	return a.checkDuplicateTxIdsImpl(txID, recentIds)
-}
-
-type appendBlockParams struct {
-	transactions   []proto.Transaction
-	chans          *verifierChans
-	block, parent  *proto.BlockHeader
-	height         uint64
-	initialisation bool
-}
-
-func (a *txAppender) hasAccountVerifyScript(tx proto.Transaction, initialisation bool) (bool, error) {
-	senderAddr, err := proto.NewAddressFromPublicKey(a.settings.AddressSchemeCharacter, tx.GetSenderPK())
-	if err != nil {
-		return false, err
-	}
-	return a.stor.scriptsStorage.newestAccountHasVerifier(senderAddr, !initialisation)
-}
-
-func (a *txAppender) orderIsScripted(order proto.Order, initialisation bool) (bool, error) {
-	return a.txHandler.tc.orderScriptedAccount(order, initialisation)
-}
-
-func (a *txAppender) handleExchange(tx proto.Transaction, blockInfo *proto.BlockInfo, initialisation bool) (uint64, error) {
-	// Smart account trading.
-	activated, err := a.stor.features.isActivated(int16(settings.SmartAccountTrading))
-	if err != nil {
-		return 0, err
-	}
-	if !activated {
-		// Functionality is not yet activated.
-		return 0, nil
-	}
-	exchange, ok := tx.(proto.Exchange)
-	if !ok {
-		return 0, errors.New("failed to convert tx to Exchange")
-	}
-	bo := exchange.GetBuyOrderFull()
-	so := exchange.GetSellOrderFull()
-	boScripted, err := a.orderIsScripted(bo, initialisation)
-	if err != nil {
-		return 0, err
-	}
-	soScripted, err := a.orderIsScripted(so, initialisation)
-	if err != nil {
-		return 0, err
-	}
-	scriptsRuns := uint64(0)
-	if boScripted {
-		if err := a.sc.callAccountScriptWithOrder(bo, blockInfo, initialisation); err != nil {
-			return 0, errors.Errorf("BUY ORDER: callAccountScriptWithOrder(): %v", err)
-		}
-		scriptsRuns++
-	}
-	if soScripted {
-		if err := a.sc.callAccountScriptWithOrder(so, blockInfo, initialisation); err != nil {
-			return 0, errors.Errorf("SELL ORDER: callAccountScriptWithOrder(): %v", err)
-		}
-		scriptsRuns++
-	}
-	activated, err = a.stor.features.isActivated(int16(settings.Ride4DApps))
-	if err != nil {
-		return 0, err
-	}
-	if !activated {
-		// Don't count before Ride4DApps activation.
-		scriptsRuns = 0
-	}
-	return scriptsRuns, nil
-}
-
-// For UTX validation, this returns the last stable block, which is in fact
-// current block.
-// For appendBlock(), this returns block that is currently being added.
-func (a *txAppender) currentBlock() (*proto.BlockHeader, error) {
-	curBlockHeight, err := a.state.AddingBlockHeight()
-	if err != nil {
-		return nil, err
-	}
-	curHeader, err := a.state.NewestHeaderByHeight(curBlockHeight)
-	if err != nil {
-		return nil, err
-	}
-	return curHeader, nil
-}
-
-func (a *txAppender) currentBlockInfo() (*proto.BlockInfo, error) {
-	height, err := a.state.AddingBlockHeight()
-	if err != nil {
-		return nil, err
-	}
-	curHeader, err := a.currentBlock()
-	if err != nil {
-		return nil, err
-	}
-	return proto.BlockInfoFromHeader(a.settings.AddressSchemeCharacter, curHeader, height)
-}
-
-func (a *txAppender) checkTxAgainstState(tx proto.Transaction, accountScripted bool, checkerInfo *checkerInfo, blockInfo *proto.BlockInfo) (uint64, error) {
-	scriptsRuns := uint64(0)
-	if accountScripted {
-		// Check script.
-		if err := a.sc.callAccountScriptWithTx(tx, blockInfo, checkerInfo.initialisation); err != nil {
-			return 0, errors.Errorf("callAccountScriptWithTx(): %v", err)
-		}
-		scriptsRuns++
-	}
-	// Check against state.
-	txSmartAssets, err := a.txHandler.checkTx(tx, checkerInfo)
-	if err != nil {
-		return 0, err
-	}
-	for _, smartAsset := range txSmartAssets {
-		if tx.GetTypeVersion().Type == proto.SetAssetScriptTransaction {
-			// Exception: don't count before Ride4DApps activation.
-			break
-		}
-		// Check smart asset's script.
-		if err := a.sc.callAssetScript(tx, smartAsset, blockInfo, checkerInfo.initialisation); err != nil {
-			return 0, errors.Errorf("callAssetScript(): %v", err)
-		}
-		scriptsRuns++
-	}
-	if tx.GetTypeVersion().Type == proto.ExchangeTransaction {
-		exchangeScripsRuns, err := a.handleExchange(tx, blockInfo, checkerInfo.initialisation)
-		if err != nil {
-			return 0, errors.Errorf("failed to handle exchange tx: %v", err)
-		}
-		scriptsRuns += exchangeScripsRuns
-	}
-	return scriptsRuns, nil
-}
-
-func (a *txAppender) checkScriptsLimits(scriptsRuns uint64) error {
-	smartAccountsActivated, err := a.stor.features.isActivated(int16(settings.SmartAccounts))
-	if err != nil {
-		return err
-	}
-	ride4DAppsActivated, err := a.stor.features.isActivated(int16(settings.Ride4DApps))
-	if err != nil {
-		return err
-	}
-	if ride4DAppsActivated {
-		if a.sc.getTotalComplexity() > maxScriptsComplexityInBlock {
-			zap.S().Warn("complexity limit per block is exceeded")
-		}
-		return nil
-	} else if smartAccountsActivated {
-		if scriptsRuns > maxScriptsRunsInBlock {
-			return errors.Errorf("more scripts runs in block than allowed: %d > %d", scriptsRuns, maxScriptsRunsInBlock)
-		}
-	}
-	return nil
-}
-
-func (a *txAppender) needToCheckOrdersSigs(transaction proto.Transaction, initialisation bool) (bool, bool, error) {
-	tx, ok := transaction.(proto.Exchange)
-	if !ok {
-		return false, false, nil
-	}
-	soScripted, err := a.orderIsScripted(tx.GetSellOrderFull(), initialisation)
-	if err != nil {
-		return false, false, err
-	}
-	boScripted, err := a.orderIsScripted(tx.GetBuyOrderFull(), initialisation)
-	if err != nil {
-		return false, false, err
-	}
-	return !soScripted, !boScripted, nil
-}
-
-func (a *txAppender) appendBlock(params *appendBlockParams) error {
-	hasParent := params.parent != nil
-	// Create miner balance diff.
-	// This adds 60% of prev block fees as very first balance diff of the current block
-	// in case NG is activated, or empty diff otherwise.
-	minerDiff, err := a.blockDiffer.createMinerDiff(params.block, hasParent, params.height)
-	if err != nil {
-		return err
-	}
-	// Save miner diff first.
-	if err := a.diffStor.saveTxDiff(minerDiff); err != nil {
-		return err
-	}
-	curHeight := params.height + 1
-	scriptsRuns := uint64(0)
-	for _, tx := range params.transactions {
-		// Detect what signatures must be checked for this transaction.
-		senderAddr, err := proto.NewAddressFromPublicKey(a.settings.AddressSchemeCharacter, tx.GetSenderPK())
-		if err != nil {
-			return err
-		}
-		accountHasVerifierScript, err := a.stor.scriptsStorage.newestAccountHasVerifier(senderAddr, !params.initialisation)
-		if err != nil {
-			return err
-		}
-		checkTxSig := true
-		if accountHasVerifierScript {
-			// For transaction with SmartAccount we don't check signatures.
-			checkTxSig = false
-		}
-		checkSellOrder, checkBuyOrder, err := a.needToCheckOrdersSigs(tx, params.initialisation)
-		if err != nil {
-			return err
-		}
-		// Send transaction for signature/data verification.
-		task := &verifyTask{
-			taskType:       verifyTx,
-			tx:             tx,
-			checkTxSig:     checkTxSig,
-			checkSellOrder: checkSellOrder,
-			checkBuyOrder:  checkBuyOrder,
-		}
-		select {
-		case verifyError := <-params.chans.errChan:
-			return verifyError
-		case params.chans.tasksChan <- task:
-		}
-		checkerInfo := &checkerInfo{
-			initialisation:   params.initialisation,
-			currentTimestamp: params.block.Timestamp,
-			blockID:          params.block.BlockSignature,
-			blockVersion:     params.block.Version,
-			height:           params.height,
-		}
-		if hasParent {
-			checkerInfo.parentTimestamp = params.parent.Timestamp
-		}
-		if err := a.checkDuplicateTxIds(tx, a.recentTxIds, params.block.Timestamp); err != nil {
-			return err
-		}
-		// Add transaction ID.
-		txID, err := tx.GetID()
-		if err != nil {
-			return err
-		}
-		a.recentTxIds[string(txID)] = empty
-		// Check against state.
-		blockInfo, err := a.currentBlockInfo()
-		if err != nil {
-			return err
-		}
-		txScriptsRuns, err := a.checkTxAgainstState(tx, accountHasVerifierScript, checkerInfo, blockInfo)
-		if err != nil {
-			return err
-		}
-		scriptsRuns += txScriptsRuns
-		if tx.GetTypeVersion().Type == proto.InvokeScriptTransaction {
-			// Invoke is handled in a special way.
-			invokeTx, ok := tx.(*proto.InvokeScriptV1)
-			if !ok {
-				return errors.New("failed to convert InvokeScriptTransaction to type InvokeScriptV1")
-			}
-			invokeInfo := &invokeAddlInfo{
-				previousScriptRuns: txScriptsRuns,
-				initialisation:     params.initialisation,
-				block:              params.block,
-				height:             curHeight,
-				validatingUtx:      false,
-			}
-			if err := a.ia.applyInvokeScriptV1(invokeTx, invokeInfo); err != nil {
-				return errors.Errorf("failed to apply InvokeScript transaction %s to state: %v", invokeTx.ID.String(), err)
-			}
-		} else {
-			// Create balance diff of this tx.
-			txDiff, err := a.blockDiffer.createTransactionDiff(tx, params.block, curHeight, params.initialisation)
-			if err != nil {
-				return err
-			}
-			// Save balance diff of this tx.
-			if err := a.diffStor.saveTxDiff(txDiff); err != nil {
-				return err
-			}
-		}
-		// Count current tx fee.
-		if err := a.blockDiffer.countMinerFee(tx); err != nil {
-			return err
-		}
-		// Perform state changes.
-		performerInfo := &performerInfo{
-			initialisation: params.initialisation,
-			blockID:        params.block.BlockSignature,
-		}
-		if err := a.txHandler.performTx(tx, performerInfo); err != nil {
-			return err
-		}
-		// Save transaction bytes to storage.
-		// TODO: not all transactions implement WriteTo.
-		bts, err := tx.MarshalBinary()
-		if err != nil {
-			return err
-		}
-		if err := a.rw.writeTransaction(txID, bts); err != nil {
-			return err
-		}
-	}
-	if err := a.checkScriptsLimits(scriptsRuns); err != nil {
-		return errors.Errorf("%s: %v", params.block.BlockSignature.String(), err)
-	}
-	// Reset block complexity counter.
-	a.sc.resetComplexity()
-	// Save fee distribution of this block.
-	// This will be needed for createMinerDiff() of next block due to NG.
-	if err := a.blockDiffer.saveCurFeeDistr(params.block); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (a *txAppender) applyAllDiffs(initialisation bool) error {
-	changes := a.diffStor.allChanges()
-	a.recentTxIds = make(map[string]struct{})
-	a.diffStor.reset()
-	if err := a.diffApplier.applyBalancesChanges(changes, !initialisation); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (a *txAppender) checkUtxTxSig(tx proto.Transaction, scripted bool) error {
-	// Check tx signature and data.
-	checkSellOrder, checkBuyOrder, err := a.needToCheckOrdersSigs(tx, false)
-	if err != nil {
-		return err
-	}
-	if err := checkTx(tx, !scripted, checkSellOrder, checkBuyOrder); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (a *txAppender) handleInvoke(tx proto.Transaction, height uint64, block *proto.BlockHeader, prevScriptRuns uint64) error {
-	invokeTx, ok := tx.(*proto.InvokeScriptV1)
-	if !ok {
-		return errors.New("failed to convert transaction to type InvokeScriptV1")
-	}
-	invokeInfo := &invokeAddlInfo{
-		previousScriptRuns: prevScriptRuns,
-		initialisation:     false,
-		block:              block,
-		height:             height,
-		validatingUtx:      true,
-	}
-	if err := a.ia.applyInvokeScriptV1(invokeTx, invokeInfo); err != nil {
-		return errors.Wrap(err, "InvokeScript validation failed")
-	}
-	return nil
-}
-
-func (a *txAppender) resetValidationList() {
-	a.sc.resetComplexity()
-	a.totalScriptsRuns = 0
-	a.recentTxIds = make(map[string]struct{})
-	a.diffStor.reset()
-}
-
-// For UTX validation.
-func (a *txAppender) validateNextTx(tx proto.Transaction, currentTimestamp, parentTimestamp uint64, version proto.BlockVersion) error {
-	if err := a.checkDuplicateTxIds(tx, a.recentTxIds, currentTimestamp); err != nil {
-		return err
-	}
-	// Add transaction ID.
-	txID, err := tx.GetID()
-	if err != nil {
-		return err
-	}
-	a.recentTxIds[string(txID)] = empty
-	scripted, err := a.hasAccountVerifyScript(tx, false)
-	if err != nil {
-		return err
-	}
-	// Check tx signature and data.
-	if err := a.checkUtxTxSig(tx, scripted); err != nil {
-		return err
-	}
-	// Check tx data against state.
-	height, err := a.state.AddingBlockHeight()
-	if err != nil {
-		return err
-	}
-	checkerInfo := &checkerInfo{
-		initialisation:   false,
-		currentTimestamp: currentTimestamp,
-		parentTimestamp:  parentTimestamp,
-		blockVersion:     version,
-		height:           height,
-	}
-	// TODO: Doesn't work correctly if miner doesn't work in NG mode.
-	// In this case it returns the last block instead of what is being mined.
-	block, err := a.currentBlock()
-	if err != nil {
-		return err
-	}
-	blockInfo, err := proto.BlockInfoFromHeader(a.settings.AddressSchemeCharacter, block, height)
-	if err != nil {
-		return err
-	}
-
-	txScriptsRuns, err := a.checkTxAgainstState(tx, scripted, checkerInfo, blockInfo)
-	if err != nil {
-		return err
-	}
-	if err := a.checkScriptsLimits(a.totalScriptsRuns + txScriptsRuns); err != nil {
-		return err
-	}
-	a.totalScriptsRuns += txScriptsRuns
-	if tx.GetTypeVersion().Type == proto.InvokeScriptTransaction {
-		// Invoke is handled in a special way.
-		return a.handleInvoke(tx, height, block, txScriptsRuns)
-	}
-	// Create, validate and save balance diff.
-	diff, err := a.txHandler.createDiffTx(tx, &differInfo{
-		initialisation: false,
-		blockInfo:      &proto.BlockInfo{Timestamp: currentTimestamp},
-	})
-	if err != nil {
-		return err
-	}
-	changes, err := a.diffStor.changesByTxDiff(diff)
-	if err != nil {
-		return err
-	}
-	if err := a.diffApplier.validateBalancesChanges(changes, true); err != nil {
-		return err
-	}
-	if err := a.diffStor.saveBalanceChanges(changes); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (a *txAppender) reset() {
-	a.sc.resetComplexity()
-	a.totalScriptsRuns = 0
-	a.recentTxIds = make(map[string]struct{})
-	a.diffStor.reset()
-	a.blockDiffer.reset()
 }
 
 type stateManager struct {
